2. post를 이용한 회원가입과 이미 있는 이메일인지 검사, 데이터베이스 연동은 아직 안 되있음
