--- src/routes/join,route.ts
@@ -1,8 +1,0 @@
-import express from 'express';
-import { joinController } from '../controllers/join,controller';
-
-const router = express.Router();
-
-router.get('/',joinController);
-
-export default router;
