+++ src/routes/join.route.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import { joinController } from '../controllers/join.controller';
+
+const router = express.Router();
+
+router.post('/',joinController);
+
+export default router;
