+++ src/types/express/index.d.ts
@@ -5,6 +5,7 @@ declare global {
     interface Request {
       user?: JwtPayload & {
         id: number;
+    
         email: string;
       };
     }
